pyproject.toml
[build-system]
requires = ['poetry-core>=1.2.0', 'wheel',]
build-backend = 'poetry.core.masonry.api'

[tool.poetry]
name = 'trestlebot'
version = '0.1.0'
description = "trestle-bot assists users in leveraging Compliance-Trestle in automated workflows of for OSCAL formatted compliance content management."

authors = ["Jennifer Power <jpower@redhat.com>",]

include = ['LICENSE']
license = 'Apache-2.0'
readme = 'README.md'

repository = 'https://github.com/RedHatProductSecurity/trestle-bot'

[tool.poetry.scripts]
trestle-bot = "trestlebot.cli:run"

[tool.poetry.dependencies]
python = '^3.8.1'
gitpython = "^3.1.31"
compliance-trestle = "^2.1.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.3.0"
isort = "^5.12.0"

[tool.poetry.group.tests.dependencies]
pytest = "^7.3.2"
pytest-cov = "^4.1.0"

[tool.coverage.run]
branch = true
relative_files = true
omit = [
    'tests/*',
]

[tool.pytest.ini_options]
minversion = '6.0'
addopts = """
--doctest-modules \
--cov=./ \
--cov-append \
--cov-report html:tests/reports/coverage-html \
--cov-report xml:tests/reports/coverage.xml \
--ignore=docs/
"""
testpaths = [
    'tests',
]
